fix: update timeDimensions metadata after actions that change them
