+++ src/components/MainSidebar/TimeDimensions.js
@@ -1,4 +1,3 @@
-import { DIMENSION_ID_PERIOD } from '@dhis2/analytics'
 import i18n from '@dhis2/d2-i18n'
 import React from 'react'
 import { useSelector } from 'react-redux'
@@ -9,76 +8,47 @@ import {
     DIMENSION_TYPE_SCHEDULED_DATE,
     DIMENSION_TYPE_LAST_UPDATED,
 } from '../../modules/dimensionTypes.js'
-import {
-    getTimeDimensions,
-    NAME_PARENT_PROPERTY_PROGRAM,
-} from '../../modules/timeDimensions.js'
-import { OUTPUT_TYPE_EVENT } from '../../modules/visualization.js'
 import { sGetMetadataById } from '../../reducers/metadata.js'
-import {
-    sGetUiInputType,
-    sGetUiProgramId,
-    sGetUiProgramStageId,
-} from '../../reducers/ui.js'
 import { DimensionItem } from './DimensionItem/index.js'
 import { MainSidebarSection } from './MainSidebarSection.js'
-import { PROGRAM_TYPE_WITH_REGISTRATION } from './ProgramDimensionsPanel/ProgramDimensionsPanel.js'
 import { useSelectedDimensions } from './SelectedDimensionsContext.js'
-
-const getName = (dimension, program, stage) => {
-    if (!dimension.nameParentProperty) {
-        return dimension.defaultName
-    }
-    const name =
-        dimension.nameParentProperty === NAME_PARENT_PROPERTY_PROGRAM
-            ? program[dimension.nameProperty]
-            : stage[dimension.nameProperty]
-    return name || dimension.defaultName
-}
 
 export const TimeDimensions = () => {
     const { getIsDimensionSelected } = useSelectedDimensions()
-    const selectedInputType = useSelector(sGetUiInputType)
-    const programId = useSelector(sGetUiProgramId)
-    const stageId = useSelector(sGetUiProgramStageId)
-    const program =
-        useSelector((state) => sGetMetadataById(state, programId)) || {}
-    const stage = useSelector((state) => sGetMetadataById(state, stageId)) || {}
+    const eventDateDim = useSelector((state) =>
+        sGetMetadataById(state, DIMENSION_TYPE_EVENT_DATE)
+    )
+    const enrollmentDateDim = useSelector((state) =>
+        sGetMetadataById(state, DIMENSION_TYPE_ENROLLMENT_DATE)
+    )
+    const incidentDateDim = useSelector((state) =>
+        sGetMetadataById(state, DIMENSION_TYPE_INCIDENT_DATE)
+    )
+    const scheduledDateDim = useSelector((state) =>
+        sGetMetadataById(state, DIMENSION_TYPE_SCHEDULED_DATE)
+    )
+    const lastUpdatedDim = useSelector((state) =>
+        sGetMetadataById(state, DIMENSION_TYPE_LAST_UPDATED)
+    )
 
-    const timeDimensions = getTimeDimensions()
-    const dimensionIds = Object.keys(timeDimensions)
-    const enabledDimensionIds = []
-
-    if (selectedInputType && program.programType && stage.id) {
-        const isEvent = selectedInputType === OUTPUT_TYPE_EVENT
-        const withRegistration =
-            program.programType === PROGRAM_TYPE_WITH_REGISTRATION
-
-        if (isEvent) {
-            enabledDimensionIds.push(DIMENSION_TYPE_EVENT_DATE)
-        }
-
-        if (withRegistration) {
-            enabledDimensionIds.push(DIMENSION_TYPE_ENROLLMENT_DATE)
-
-            isEvent &&
-                !stage.hideDueDate &&
-                enabledDimensionIds.push(DIMENSION_TYPE_SCHEDULED_DATE)
-
-            program.displayIncidentDate &&
-                enabledDimensionIds.push(DIMENSION_TYPE_INCIDENT_DATE)
-        }
-
-        if (isEvent || withRegistration) {
-            enabledDimensionIds.push(DIMENSION_TYPE_LAST_UPDATED)
-        }
+    if (!eventDateDim) {
+        return null
     }
 
-    const dimensions = dimensionIds.map((dimensionId) => ({
-        id: dimensionId,
-        dimensionType: DIMENSION_ID_PERIOD,
-        name: getName(timeDimensions[dimensionId], program, stage),
-        isDisabled: !enabledDimensionIds.includes(dimensionId),
+    const dimensions = [
+        eventDateDim,
+        enrollmentDateDim,
+        incidentDateDim,
+        scheduledDateDim,
+        lastUpdatedDim,
+    ].map(({ id, name, dimensionType, disabled }) => ({
+        id,
+        dimensionType,
+        name,
+        selected: getIsDimensionSelected(id),
+        disabled,
+        //optionSet
+        //valueType
     }))
 
     return (
@@ -90,9 +60,7 @@ export const TimeDimensions = () => {
                     name={dimension.name}
                     id={dimension.id}
                     selected={getIsDimensionSelected(dimension.id)}
-                    disabled={dimension.isDisabled}
-                    optionSet={dimension.optionSet}
-                    valueType={dimension.valueType}
+                    disabled={dimension.disabled}
                 />
             ))}
         </MainSidebarSection>
